Update runLatex.sh
idem

+++ img/recetteCookbook/addText2Img/runLatex.sh
@@ -1,3 +1,7 @@
-rm img/kardinalschnittenSchema.pdf
+set -e
+set -o pipefail
+DIRECTORY=`dirname $0`
+cd $DIRECTORY
+[ -e  img/kardinalschnittenSchema.pdf ] && rm img/kardinalschnittenSchema.pdf
 pdflatex --shell-escape  main.tex
 ln -fs addText2Img/img/kardinalschnittenSchema.pdf ../
